ARPA-50 fixed existing tests
