+++ Tests/Orso.Arpa.Domain.Tests/AuthTests/CommandHandlerTests/SetRoleHandlerTests.cs
@@ -1,7 +1,6 @@
 using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
-using MediatR;
 using NUnit.Framework;
 using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Domain.Identity;
@@ -36,10 +35,10 @@ namespace Orso.Arpa.Domain.Tests.AuthTests.CommandHandlerTests
             };
 
             // Act
-            Unit result = await _handler.Handle(command, new CancellationToken());
+            bool result = await _handler.Handle(command, new CancellationToken());
 
             // Assert
-            result.Should().BeEquivalentTo(Unit.Value);
+            result.Should().BeFalse();
         }
     }
 }
